Move client_utilities to ClientUtilities and client_exceptions to ClientExceptions. Create UserUtilities module with user exposed methods (move user exposed methods from ClientUtilities to UserUtilities). Still keep a client_utilities module with the user exposed methods from UserUtilities imported, and printing a module retirement message. Fixes #4358.
