+++ src/python/CRABClient/Commands/getcommand.py
@@ -1,7 +1,7 @@
 from CRABClient.Commands.remote_copy import remote_copy
 from CRABClient.Commands.SubCommand import SubCommand
-from CRABClient.client_exceptions import ConfigurationException , RESTCommunicationException
-from CRABClient.client_utilities import validateJobids
+from CRABClient.ClientExceptions import ConfigurationException , RESTCommunicationException
+from CRABClient.ClientUtilities import validateJobids
 from CRABClient import __version__
 import CRABClient.Emulator
