+++ src/python/CRABClient/Commands/getlog.py
@@ -1,7 +1,7 @@
 from CRABClient.Commands.remote_copy import remote_copy
 from CRABClient.Commands.SubCommand import SubCommand
 from CRABClient.Commands.getcommand import getcommand
-from CRABClient.client_exceptions import ConfigurationException
+from CRABClient.ClientExceptions import ConfigurationException
 
 import os
